Use uint32_t for protocol-defined size fields
- udx_value_entry_item.size: size_t -> uint32_t (matches value_size in protocol)
- udx_db_builder_create_with_metadata metadata_size: size_t -> uint32_t
  (matches metadata_size in db header)
- Remove redundant UINT32_MAX overflow check and cast

+++ src/udx_writer.h
@@ -76,7 +76,7 @@ udx_db_builder *udx_db_builder_create(udx_writer *writer, const char *name);
 udx_db_builder *udx_db_builder_create_with_metadata(udx_writer *writer,
                                                     const char *name,
                                                     const uint8_t *metadata,
-                                                    size_t metadata_size);
+                                                    uint32_t metadata_size);
 
 /**
  * Finalize building the database
